chore: extract modal content to separate component
