+++ dev/vite/index.html
@@ -88,6 +88,11 @@
     <h3>Icon only</h3>
     <lwc-button icon-only></lwc-button>
 
+    <h2>Modal Content</h2>
+    <div style="width: 448px">
+      <lwc-modal-content></lwc-modal-content>
+    </div>
+
     <h2>Connector List</h2>
     <lwc-connector-list></lwc-connector-list>
   </body>
